add plugins as submodules
