+++ templates/layouts/main.html
@@ -51,6 +51,9 @@
     <script src="js/leaflet.fusesearch.js"></script>
     <link href="css/leaflet.fusesearch.css" rel="stylesheet">
 
+    <!-- Custom override of leaflet.fusesearch.js -->
+    <script src="js/leaflet.fusesearch.extended.js"></script>
+
     <!-- Custom CSS -->
     <link href="css/stac-ui.css" rel="stylesheet">
     <link href="css/leaflet-ui.css" rel="stylesheet">
